[playground] Rewrite using `iframe`
